+++ cardano-db-sync/src/Cardano/DbSync/Era/Shelley/Generic/Tx/Alonzo.hs
@@ -40,6 +40,7 @@ import           Cardano.DbSync.Era.Shelley.Generic.Witness
 
 import qualified Cardano.Ledger.Address as Ledger
 import qualified Cardano.Ledger.Alonzo.Data as Alonzo
+import qualified Cardano.Ledger.Alonzo.Language as Alonzo
 import           Cardano.Ledger.Alonzo.Scripts (ExUnits (..), txscriptfee)
 import qualified Cardano.Ledger.Alonzo.Scripts as Alonzo
 import           Cardano.Ledger.Alonzo.Tx (ValidatedTx (..))
@@ -276,7 +277,8 @@ mkTxScript (hsh, script) =
     getScriptType =
       case script of
         Alonzo.TimelockScript {} -> Timelock
-        Alonzo.PlutusScript {} -> Plutus
+        Alonzo.PlutusScript Alonzo.PlutusV1 _s -> PlutusV1
+        Alonzo.PlutusScript Alonzo.PlutusV2 _s -> PlutusV2
 
     timelockJsonScript :: Maybe ByteString
     timelockJsonScript =
@@ -289,8 +291,12 @@ mkTxScript (hsh, script) =
     plutusCborScript =
       case script of
         Alonzo.TimelockScript {} -> Nothing
-        Alonzo.PlutusScript _lang s ->
+        -- The serialization in cardano-api ignores the script version tag,
+        -- but we handle both cases properly for posterity.
+        Alonzo.PlutusScript Alonzo.PlutusV1 s ->
           Just . Api.serialiseToCBOR . Api.PlutusScript Api.PlutusScriptV1 $ Api.PlutusScriptSerialised s
+        Alonzo.PlutusScript Alonzo.PlutusV2 s ->
+          Just . Api.serialiseToCBOR . Api.PlutusScript Api.PlutusScriptV2 $ Api.PlutusScriptSerialised s
 
 getPlutusSizes ::
     ( HasField "wits" tx (Alonzo.TxWitness era)